refactored the GET router to assignment-service getAssignments function
