+++ package.json
@@ -59,7 +59,7 @@
     "input-otp": "^1.4.2",
     "lucide-react": "^0.454.0",
     "next": "15.5.9",
-    "next-auth": "4.24.7",
+    "next-auth": "^4.24.14",
     "next-themes": "^0.4.4",
     "nodemailer": "^6.9.13",
     "prisma": "^6.19.2",
